+++ Pod/Classes/Theme.swift
@@ -23,7 +23,7 @@ import Foundation
 
 #endif
 
-private typealias RPThemeDict = [String:[String:AnyObject]]
+private typealias RPThemeDict = [String: [AnyHashable: AnyObject]]
 private typealias RPThemeStringDict = [String:[String:String]]
 
 /// Theme parser, can be used to configure the theme parameters. 
@@ -38,8 +38,8 @@ open class Theme {
     /// Italic font to be used by this theme
     open var italicCodeFont : RPFont!
     
-    fileprivate var themeDict : RPThemeDict!
-    fileprivate var strippedTheme : RPThemeStringDict!
+    private var themeDict : RPThemeDict!
+    private var strippedTheme : RPThemeStringDict!
     
     /// Default background color for the current theme.
     open var themeBackgroundColor : RPColor!
@@ -72,7 +72,7 @@ open class Theme {
             }else
             {
                 let range = bkgColorHex.range(of: "#")
-                let str = bkgColorHex.substring(from: (range?.lowerBound)!)
+                let str = String(bkgColorHex[(range?.lowerBound)!...])
                 themeBackgroundColor = colorWithHexString(str)
             }
         }else
@@ -91,19 +91,19 @@ open class Theme {
         codeFont = font
         
         #if os(iOS) || os(tvOS)
-        let boldDescriptor = UIFontDescriptor(fontAttributes: [UIFontDescriptorFamilyAttribute:font.familyName,
-                                                                UIFontDescriptorFaceAttribute:"Bold"])
-        let italicDescriptor = UIFontDescriptor(fontAttributes: [UIFontDescriptorFamilyAttribute:font.familyName,
-                                                                 UIFontDescriptorFaceAttribute:"Italic"])
-        let obliqueDescriptor = UIFontDescriptor(fontAttributes: [UIFontDescriptorFamilyAttribute:font.familyName,
-                                                                  UIFontDescriptorFaceAttribute:"Oblique"])
+        let boldDescriptor = UIFontDescriptor(fontAttributes: [UIFontDescriptor.AttributeName.family:font.familyName,
+                                                                UIFontDescriptor.AttributeName.face:"Bold"])
+        let italicDescriptor = UIFontDescriptor(fontAttributes: [UIFontDescriptor.AttributeName.family:font.familyName,
+                                                                 UIFontDescriptor.AttributeName.face:"Italic"])
+        let obliqueDescriptor = UIFontDescriptor(fontAttributes: [UIFontDescriptor.AttributeName.family:font.familyName,
+                                                                  UIFontDescriptor.AttributeName.face:"Oblique"])
         #else
-        let boldDescriptor = NSFontDescriptor(fontAttributes: [NSFontFamilyAttribute:font.familyName!,
-                                                                NSFontFaceAttribute:"Bold"])
-        let italicDescriptor = NSFontDescriptor(fontAttributes: [NSFontFamilyAttribute:font.familyName!,
-                                                                NSFontFaceAttribute:"Italic"])
-        let obliqueDescriptor = NSFontDescriptor(fontAttributes: [NSFontFamilyAttribute:font.familyName!,
-                                                                NSFontFaceAttribute:"Oblique"])
+        let boldDescriptor = NSFontDescriptor(fontAttributes: [.family:font.familyName!,
+                                                                   .face:"Bold"])
+        let italicDescriptor = NSFontDescriptor(fontAttributes: [.family:font.familyName!,
+                                                                     .face:"Italic"])
+        let obliqueDescriptor = NSFontDescriptor(fontAttributes: [.family:font.familyName!,
+                                                                      .face:"Oblique"])
         #endif
         
         boldCodeFont = RPFont(descriptor: boldDescriptor, size: font.pointSize)
@@ -134,11 +134,11 @@ open class Theme {
         
         if styleList.count > 0
         {
-            var attrs = [String:AnyObject]()
-            attrs[NSFontAttributeName] = codeFont
+            var attrs = [NSAttributedStringKey: Any]()
+            attrs[.font] = codeFont
             for style in styleList
             {
-                if let themeStyle = themeDict[style]
+                if let themeStyle = themeDict[style] as? [NSAttributedStringKey: Any]
                 {
                     for (attrName, attrValue) in themeStyle
                     {
@@ -151,13 +151,13 @@ open class Theme {
         }
         else
         {
-            returnString = NSAttributedString(string: string, attributes:[NSFontAttributeName:codeFont] )
+            returnString = NSAttributedString(string: string, attributes:[NSAttributedStringKey.font:codeFont] )
         }
         
         return returnString
     }
     
-    fileprivate func stripTheme(_ themeString : String) -> [String:[String:String]]
+    private func stripTheme(_ themeString : String) -> [String:[String:String]]
     {
         let objcString = (themeString as NSString)
         let cssRegex = try! NSRegularExpression(pattern: "(?:(\\.[a-zA-Z0-9\\-_]*(?:[, ]\\.[a-zA-Z0-9\\-_]*)*)\\{([^\\}]*?)\\})", options:[.caseInsensitive])
@@ -172,7 +172,7 @@ open class Theme {
             if(result.numberOfRanges == 3)
             {
                 var attributes = [String:String]()
-                let cssPairs = objcString.substring(with: result.rangeAt(2)).components(separatedBy: ";")
+                let cssPairs = objcString.substring(with: result.range(at: 2)).components(separatedBy: ";")
                 for pair in cssPairs {
                     let cssPropComp = pair.components(separatedBy: ":")
                     if(cssPropComp.count == 2)
@@ -183,7 +183,7 @@ open class Theme {
                 }
                 if attributes.count > 0
                 {
-                    resultDict[objcString.substring(with: result.rangeAt(1))] = attributes
+                    resultDict[objcString.substring(with: result.range(at: 1))] = attributes
                 }
                 
             }
@@ -213,7 +213,7 @@ open class Theme {
         return returnDict
     }
     
-    fileprivate func strippedThemeToString(_ theme: RPThemeStringDict) -> String
+    private func strippedThemeToString(_ theme: RPThemeStringDict) -> String
     {
         var resultString = ""
         for (key, props) in theme {
@@ -230,12 +230,12 @@ open class Theme {
         return resultString
     }
     
-    fileprivate func strippedThemeToTheme(_ theme: RPThemeStringDict) -> RPThemeDict
+    private func strippedThemeToTheme(_ theme: RPThemeStringDict) -> RPThemeDict
     {
         var returnTheme = RPThemeDict()
         for (className, props) in theme
         {
-            var keyProps = [String:AnyObject]()
+            var keyProps = [NSAttributedStringKey: AnyObject]()
             for (key, prop) in props
             {
                 switch key
@@ -265,7 +265,7 @@ open class Theme {
         return returnTheme
     }
     
-    fileprivate func fontForCSSStyle(_ fontStyle:String) -> RPFont
+    private func fontForCSSStyle(_ fontStyle:String) -> RPFont
     {
         switch fontStyle
         {
@@ -278,23 +278,23 @@ open class Theme {
         }
     }
     
-    fileprivate func attributeForCSSKey(_ key: String) -> String
+    private func attributeForCSSKey(_ key: String) -> NSAttributedStringKey
     {
         switch key {
         case "color":
-            return NSForegroundColorAttributeName
+            return .foregroundColor
         case "font-weight":
-            return NSFontAttributeName
+            return .font
         case "font-style":
-            return NSFontAttributeName
+            return .font
         case "background-color":
-            return NSBackgroundColorAttributeName
+            return .backgroundColor
         default:
-            return NSFontAttributeName
+            return .font
         }
     }
     
-    fileprivate func colorWithHexString (_ hex:String) -> RPColor
+    private func colorWithHexString (_ hex:String) -> RPColor
     {
 
         var cString:String = hex.trimmingCharacters(in: CharacterSet.whitespacesAndNewlines)
@@ -321,7 +321,7 @@ open class Theme {
             }
         }
         
-        if (cString.characters.count != 6 && cString.characters.count != 3 )
+        if (cString.count != 6 && cString.count != 3 )
         {
             return RPColor.gray
         }
@@ -330,7 +330,7 @@ open class Theme {
         var r:CUnsignedInt = 0, g:CUnsignedInt = 0, b:CUnsignedInt = 0;
         var divisor : CGFloat
         
-        if (cString.characters.count == 6 )
+        if (cString.count == 6 )
         {
         
             let rString = (cString as NSString).substring(to: 2)